1.0
First iteration of the program that reads a list of books from a set file. The list of books will then be sorted by release year and printed out to the user along with the average star rating (1-5 with a special case for 0) and the number of books unfinished (the special 0 case, considered with a state of Did Not Finish).